fixed comments, removed ExecuteOpcode method, moved table creation to instructions.go, and minor style changes
